FD-33957-publish mentions to tms (#5628)
* FD-33957-publish mentions to tms

* FD-33957-Review changes and rubocop fixes

* FD-33957-Rubocop fixes

* FD-33957-Rubocop changes

* FD-33957 - Test case and fixed issues

* FD-33957-Updated test cases and fixed bugs

* FD-33957- Update feature name to mentions_to_tms

* FD-33957- Fixing test case failures

* FD-33957-Fixing failing test case

* Review changes

* changing syntax error

* FD-33957-Fixing visibility issue

+++ app/controllers/admin/social/twitter_streams_controller.rb
@@ -77,6 +77,8 @@ class Admin::Social::TwitterStreamsController < Admin::Social::StreamsController
       update_ticket_rules
       return 
     end
+    return construct_rule_params if Account.current.mentions_to_tms_enabled?
+
     if dont_convert_tweets_to_ticket?
       @twitter_stream.ticket_rules.delete_all
     elsif rules_without_keywords?
@@ -84,9 +86,33 @@ class Admin::Social::TwitterStreamsController < Admin::Social::StreamsController
     else
       update_ticket_rules   
       update_smart_filter_rule_with_keywords if smart_filter_feature_enabled?
-    end  
+    end
   end
 
+  def construct_rule_params
+    if dont_convert_tweets_to_ticket?
+      delete_all_rules = true
+    elsif rules_without_keywords?
+      if smart_filter_feature_enabled?
+        smart_rules, deleted_rules = update_smart_filter_rule_without_keywords
+      else
+        rules, deleted_rules = update_ticket_rules
+      end
+    else
+      rules, deleted_rules = update_ticket_rules
+      smart_rules, deleted_rules_smart_filter = update_smart_filter_rule_with_keywords if smart_filter_feature_enabled?
+    end
+    stream_update_params = { accessible_attributes: stream_accessible_params(@twitter_stream) }
+    params = {
+      rules: rules,
+      smart_rules: smart_rules,
+      stream_update_params: stream_update_params,
+      delete_all_rules: delete_all_rules,
+      deleted_rules: deleted_rules,
+      deleted_rules_smart_filter: deleted_rules_smart_filter
+    }
+    @twitter_stream.update_all_rules(params)
+  end
 
   def show_error
     flash[:notice] = @ticket_error_flash
@@ -118,7 +144,7 @@ class Admin::Social::TwitterStreamsController < Admin::Social::StreamsController
   def update_twitter_handle
     twitter_handle_params = construct_handle_params
     if @twitter_handle.update_attributes(twitter_handle_params)
-       @twitter_stream.update_attributes({:accessible_attributes => stream_accessible_params(@twitter_stream)})
+      @twitter_stream.update_attributes(accessible_attributes: stream_accessible_params(@twitter_stream)) unless Account.current.mentions_to_tms_enabled?
     else
       @ticket_error_flash = t('admin.social.flash.stream_save_error')
      end
@@ -181,6 +207,7 @@ class Admin::Social::TwitterStreamsController < Admin::Social::StreamsController
   end
 
   def update_ticket_rules
+    keyword_filter_rules = []
     deleted_rules = Array.new
     params[:social_ticket_rule].each do |rule|
       group_id = rule[:group_id].to_i unless (rule[:group_id].to_i == 0)
@@ -204,13 +231,27 @@ class Admin::Social::TwitterStreamsController < Admin::Social::StreamsController
       }
       rule_params[:action_data].merge!({:convert_all => true}) if rule[:convert_all]
       if rule[:ticket_rule_id].empty?
+        if publish_mention_rules?
+          keyword_filter_rules << construct_mention_rules(rule_params, rule, 'create')
+          next
+        end
         ticket_rule = @twitter_stream.ticket_rules.new(rule_params)
         ticket_rule.save
       else
+        if publish_mention_rules?
+          keyword_filter_rules << construct_mention_rules(rule_params, rule, 'update')
+          next
+        end
         @twitter_stream.ticket_rules.find_by_id(rule[:ticket_rule_id]).update_attributes(rule_params)
       end
     end
-    delete_rules(deleted_rules.compact) 
+    return keyword_filter_rules, deleted_rules if publish_mention_rules?
+
+    delete_rules(deleted_rules.compact)
+  end
+
+  def publish_mention_rules?
+    Account.current.mentions_to_tms_enabled? && !@twitter_stream.custom_stream?
   end
 
 
@@ -220,23 +261,34 @@ class Admin::Social::TwitterStreamsController < Admin::Social::StreamsController
       smart_rule.merge!(:with_keywords => 1)
       update_smart_filter_rule(smart_rule) 
     else
-      delete_rules([@twitter_stream.smart_filter_rule])
+      sf_rule = @twitter_stream.smart_filter_rule
+      return nil, sf_rule if Account.current.mentions_to_tms_enabled?
+      
+      delete_rules([sf_rule])
     end    
   end
 
   def update_smart_filter_rule_without_keywords
-    @twitter_stream.delete_keyword_rules
+    if Account.current.mentions_to_tms_enabled?
+      @deleted_rules = @twitter_stream.keyword_rules.map(&:id)
+    else
+      @twitter_stream.delete_keyword_rules
+    end
     smart_rule = params[:smart_filter_rule_without_keywords]
     smart_rule.merge!(:with_keywords => 0)
-    update_smart_filter_rule(smart_rule) 
+    update_smart_filter_rule(smart_rule)
   end
 
   def update_smart_filter_rule(rule)
     smart_rule_params = @twitter_stream.build_smart_rule(rule)
     if rule[:ticket_rule_id].blank?
+      return construct_mention_rules(smart_rule_params, rule, 'create') if Account.current.mentions_to_tms_enabled?
+
       ticket_rule = @twitter_stream.ticket_rules.new(smart_rule_params)
       ticket_rule.save
     else
+      return construct_mention_rules(smart_rule_params, rule, 'update'), @deleted_rules if Account.current.mentions_to_tms_enabled?
+
       @twitter_stream.smart_filter_rule.update_attributes(smart_rule_params)
     end  
   end
@@ -311,4 +363,11 @@ class Admin::Social::TwitterStreamsController < Admin::Social::StreamsController
     end 
   end
 
+  def construct_mention_rules(rule_params, rule, action)
+    {
+      rule_params: rule_params,
+      rule: rule,
+      action: action
+    }
+  end
 end
